Cycles: Increase max film exposure in UI
The previous max value of 10 has been around since the initial commit of
Cycles, so its origin is uncertain. However, this value is very low when
compared to the exposure value in Color Management. If the Color
Management exposure value is `x`, then the matching value here is `2^x`.
As such, make the soft_max and max here match the equivalent values in
Color Management exposure.

Specifying large film exposure values is useful as it affects the raw
data and thus results in the same color values available in EXR, Viewer,
and Compositor, whereas using Color Management exposure is not reflected
in EXR exports, resulting in unusably low values. Using film exposure
remedies that.