UI: Move the asterisk before "*Save Preferences"
In the User Preferences, when the Auto-Save Preferences option is
disabled, a "Save Preferences" button appears. When the prefs are
modified, it gets an asterisk after the label. However, the left
sidebar is often too narrow to see the asterisk, so you need to widen
it to see whether the preferences have been changed.

To fix this issue, this commit moves the asterisk to the left side of
the text, so that it is visible regardless of the width of the
sidebar. This is also in keeping with the asterisk in the title bar.