UI: Rename "Degr" socket of the Rotate compositing node to "Angle"
The "Degr" socket of the Rotate compositing node is used to store the
angle of rotation. It is needlessly shortened, but it is also wrong,
as the unit depends on the scene settings and can also be radians.

Similarly to !108234, this commit renames it to "Angle", with a
versioning for existing nodes.