Build manifold library in extern.
I copied the header-only libraries thrust and glm into the
third_party directory of manifold and then make a simple CMakeLists.txt
that just included the needed source files with a couple defines.
This setup is useful while experimenting. It makes it easier to debug
and profile.
Also, by doing this, it was easy to remove the dependencies on hull
and clipper2 (just a few ifdefs around code we don't need anyway for
Blender boolean).