Fix: artifacts at volume self-intersection
previously, we ignore volume front face hit if the object and the
shader is the same as the last hit; this causes the ray to exit the
volume stack at the next back face hit, even if it's still in the
volume.

To fix this problem, add such front face hits to the volume stack so
that it has the correct depth for closed meshes. The density does not
accumulate at the intersection.