Fix: Jack and Coreaudio would try to sync the scene frame twice
The code tried to sync the playback to audio in two places:
1. In the screen_animation_step_invoke function in screen_ops.cc
2. In wm_event_do_handlers which would trigger on every screen refresh/event

However because of some oversight on how the Audaspace API worked,
the wm_event_do_handlers code was only triggered for JACK and CoreAudio.

Remove this code as we shouldn't need to update the current scene frame
in two places.

The wm_event_do_handlers way would also lead to bugs as it would try to
sync regardless if we were playing back audio or not.