UI: Add shadows to panels
Draw a soft shadow to panels, so they are easier to tell apart from
the background.

To avoid introducing new theme settings and since it serves a similar
purpose, use the theme menu shadow factor and width preferences
(its name has been tweaked to reflect this).

While dragging a panel, the shadow will be more prominent (if the
theme preference is below 16px) to help communicate that the panel
is drawn above others.