UI: Themeable RedAlert, Error, Warning, Info Colors
The "Red Alert" color is currently hard-coded which caused problems in
themes.  It also has a an enum value of 0, which precludes using this
value as "unset". This PR moves this to the "State" part of the theme,
and also moves "Error", "Warning", and "Info" colors there, since their
current location in the Info Editor area doesn't work well for using
these colors elsewhere.