Fix: Grease Pencil: numinput keyframes broken unless mouse move first
Oversight in 09d2108bf502

Above commit did a check for `TRANS_RUNNING` before calling
`grease_pencil_layer_update_trans_data`.
This however is not set for numinput (one could argue that it should --
e.g. in `transformEvent`, but i havent checked all usages of
`TRANS_RUNNING` in a deep enough manner that it would justify making a
change to that).

So to resolve, not only check `TRANS_RUNNING`, but also allow/call
`grease_pencil_layer_update_trans_data` if `hasNumInput` is true.

Fixes #131092