Fix #131107: Grease Pencil: Crash doing Grease Pencil Surface Reproject
Implemented in 4291ab855c30
Fixed in c20399442d5b (even though that commit introduced the crash)

The crash happens because we are dealing with multiple
`DrawingPlacement` (using the constructor from c20399442d5b which allows
for passing in cached depths) in **threads**. The depth_cache gets freed
in the destructor of  `DrawingPlacement` though, thus ruining the data
for a `DrawingPlacement` in another thread.

To resolve, keep track of how we constructed `DrawingPlacement`, if
depths were passed in these then dont get freed, the freeing then
happens "by hand" in the caller (`grease_pencil_reproject_exec`).