Fix #131194: Crash with animated geometry nodes math node operation
The animation system can't properly update the node tree to regenerate
all the necessary information when changing the operation. The operation
changing with animation also invalidates some assumptions we make to
improve performance, and the complexity of supporting it isn't worth
the questionable utility.

Similar to 7797fcc3c9edf7713878 and dc9249c97f033d492d13.