Fix #131269: Vulkan: readback of ViewDepths
The Vulkan backend misses data conversions for depth buffer when read as
GPU_DATA_UINT24_S8 for both GPU_DEPTH24_STENCIL8 and
GPU_DEPTH32F_STENCIL8. These data conversions should be added, however
where this is used in Blender requires the float values which are
supported.

This PR changes the calling code (`view3d_depths_create`) to request
the backend to do the conversion to floats.

In the future we will be migrating to using floats to store depth as it
enables us to use inverse depth which will improve depth testing.