Fix #131210: Crash on selecting brush in non-updated file
Missed in 9945d511a931e3c2f5328fd45456aed63807ca99

The previous commit fixed files that had not yet been saved in 4.3,
however, the versioning code was not bumped which means that a user who
had a file in a version prior to 4.2 that then saved with the full
release of 4.3 would not have the corrected versioning applied for their
texture paint workspaces.

This commit reorders the relevant method to the bottom of the file and
bumps the version to prevent the invalid data leading to a crash.