Fix #131106: VSE strip mask inconsistencies between byte and float images
If a strip mask was used several times in the same frame, and it
needed to do byte->float conversion (e.g. mask is produced by
a byte-color strip, but then used in both a byte-color strip, and
later on in a float-color strip), then that byte->float mask
image conversion was introducing two inconsistencies:
- It was making mask alpha channel affect the result, which was not
  happening with regular byte mask images, and
- It was converting float mask to scene linear space, which was
  resulting in different image than with a byte mask.

Fix those issues by ensuring that there's no alpha premultiplication,
and no "to scene linear" conversion of the float mask. Now both
byte and float masks used within VSE are in the same color space,
and both ignore the alpha channel.