FBX export: always write edge crease (even if no subd modifier present)
This has come up a couple of times and it seems better to change the
implementation from blender/blender-addons@f1dd37b8ac8f [which only
exported edge creases when a subd modifier with enabled `Use Creases`
option was present] in a way that these are always exported.

Also took the liberty to rename the variable name for the layer to
correspond to "crease".

Fixes #131284