Fixes #93103  Hard limits were been addded to the allowed iteration values of several modifiers to prevent undefined operations.
The affected functions are Decimate, Smooth, Corrective Smooth, Laplacian Smooth, Laplacian Deform, and Grease Pencil Simplify
Previously, the UI allowed values in the range of -32768...+32768 for each of these.  Negative numbers for "number of iterations of smooth to perform" made no sense and could lock up blender.  This prevents the user from doing that.

Negative values are now disallowed.
Zero values are still allowed -- This will result in no decimation, smoothing, deforming, simplifiying, etc being performed.

This should be compatible with existing saves with no problem.  Some modifiers quietly ignored uninterpretable (negative number of iterations) parameters and acted as though the value was zero.  Others, like smooth, misbehaved and locked up the UI, This should mean no files exist in the wild with bad parameters since it should be impossible to save such a file.

All the existing modifiers use shorts to store the number of iterations except for Laplacian Deform which stored an int.
This commit removes the user's ability to create a Laplacian Deform modifier with a number of iterations greater than 32767, which *technically* was possible before. Prior to this commit, it used to be possible to declare a number of iterations up to INT_MAX.  Allowing that was probably not intentional.

Screw is involved here too since it has an iteration count and has also been updated.  It previously allowed 1...10000 and now allows 1...INT_MAX, which is the largest representable number in the DNA for this modifier.  There could be conceivable cases where a user might want a screw with more than an arbitrary limit of 10,000 threads, and this will let the user at least attempt that.
That said, don't blame me if you try to make a 9,223,372,036,854,775,807 rotation screw and your video card starts on fire.
Annoyingly a screw iteration of 0 is (still) not permitted becuase the screw modifier makes an empty mesh in that case instead of an unmodified mesh.  Therefore the screw modifier lower bound remains at 1, not 0.