Fix #131474: Incorrect framebuffer configuration image engine overlays
When the overlay engine ius used to draw over the image engine it uses
incorrect framebuffer configurations. It can attach and sample both the
depth buffer at the same time.

This PR fixes this bu separating the output framebuffer into two. One
with and one without a depth buffer. And bind the correct one when
needed.