VSE: Change modifiers to be able to take in either byte or float masks
Some of them already were doing it, yet some others were only working
for cases of "byte image + byte mask, or float image + float mask".
Make all of them consistently be able to process either byte or float
mask, with either image type.

This makes it so that the mask image never needs to get converted
byte->float or float->byte (issue #131355), and can be just used as-is.