GPU: ensure viewport textures are cleared (#119685 and possibly many others)
GPUViewport is creating a bunch of framebuffer textures for itself,
but some space types never initialize/use them. E.g. Sequencer, Nodes
etc. only ever use the "overlay" texture. Eventually when viewport
is "drawn", it combines this uninitialized texture data and then only
by luck it happens that most of the time it is black. But not always!

The textures were only cleared (right now) on Metal backend, under
GPU_clear_viewport_workaround as if it was some driver workaround. Stop
doing that, and just clear them always.

However, there was seemingly a performance issue on OpenGL, when this
clear was being done. At least on my machine (Win10, Geforce RTX 3080Ti),
the overhead of doing the clears is measurable, and is caused by
usage of GL4.4 glClearTexImage instead of a framebuffer clear. As if
glClearTexImage makes "pixel data to exist" on the CPU side and then
later on binding this framebuffer sends off that data to the GPU, or
somesuch.

Timeline of events:
- 2.83 adds glClearTexImage code (not used yet) 738bb309
- 2.90 adds "non-glClearTexImage path is fast now" and probably
  accidentally disabled glClearTexImage usage c19cc7ef81
- 2.91 in a refactor uncomments glClearTexImage usage c766d9b9d
- 3.1 starts clearing viewport textures (just like in this commit!)
  due to many reports of random garbage due to uninitialized memory
  6738ecb
- 3.2 makes clearing only happen on Intel GPUs when on Mac or Linux
  (even if previous reports of garbage were on other GPUs), because
  now there are reports of slow UI resizing (my theory: because
  glClearTexImage incurred costs) 439f86a
- 3.2.1 makes clearing happen on Intel GPUs on Windows as well (due
  to reports of garbage) 23bce4a97
- 4.0-ish, Metal happens for MacOS, and starts to clear on all GPUs
  consistently
- 4.0 accidentally removes the "clear on all Intel GPUs" on OpenGL
  7f7e1ab08

Some bug reports start coming in about garbage rendering (mostly on
Intel GPUs which presumably clear GPU allocations much more rarely),
e.g. #119685 and #131500. But some others too, and are worked around
instead of fixing the root cause (#110976, #119512), assuming thes
are driver bugs and not Blender relying on uninitialized memory
being zeroes.

Some bug reports wrt resizing performance are still open (#97272,
#127099) with no clear pattern whether they are caused by anything
above.