Fix: use `unspecified` idroot when duplicating Action from UI
With Slotted Actions, an Action's `idroot` should always be zero,
corresponding to "unspecified". However, there was still some old
leftover code that, when duplicating an Action via the Action Editor,
would set the new Action's `idroot` to either OB or KE, depending
on the editor mode.

In addition to now being incorrect, this was also tripping an assert
elsewhere in the code when trying to then assign that duplicate action
to an ID.

This commit simply removes the outdated code, which is no longer needed,
leaving the `idroot` as the default (unspecified).