Fix #109765
This was a logic error on the "remove verts" logic.

The previous behavior tagged any vert on the mesh that was not between an edge pair.  In other words, a 3,4,5 pole, etc.
After the edge dissolve, the mesh was iterated a second time and any of those verts was checked.  IF it became an edge pair, that meant that it had been touching an edge that was removed.
This was assumed to mean that it was subject to the remove-verts logic.

However, there are times where dissolving edges will dissolve additional, non-selected edges.  This occurs becuase a face can't contain an internal vert with only one edge attached (this would require the loop to and from that edge to be tied to the same edge which can't happen.  So when it occurs, those extra edges are also removed.

When this occurs, the old behavior would say, "Well, an edge went away, so if the vert it touched is now an edge pair, we want to remove it.  This results in the unintentional removal of verts that were not part of the dissolve.

The new behavior is that instead of tagging any vert on the mesh for removal checking, the code now only tags the verts of the selected edges for possible removal.