Fix #131510: Cloth brush Persistent Base can lead to artifacts
Depending on the previous state of the mesh, the Persistent base brush
option can cause major artifacts. This is because the previous positions
and distances are used to calculate a factor when applying corrective
forces onto pairs of connected vertices.

If the distance between two vertices in the saved position is greater
than the current distance then this causes the correction vector to flip
directions, inverting the direction that the vertices should be moving.

To fix this, we clamp this distance factor to [0, 1]