Fix #131596: Grease Pencil Reproject to View is wrong
Previous method of doing the projection from the spot of the camera
(flipping the ray etc.) was not reliable (depending on where the view/
camera was located.

Instead, use the same ray origin for all methods and define the plane at
the placement location (with the view direction).
This is also easier to read I think.