Fix: version ultra-old actions properly again
The code for converting pre-Animato actions was not getting run properly
because `chanbase` was erroneously getting cleared before the relevant
versioning code was run.

The root cause was that the code checking whether an action was already
valid as a layered action or not was not confirming that `chanbase`
was empty as part of that check, which in turn triggered code that
defensively clears `chanbase` (among other things) when an action is
identified as layered.