Fix #93769: Improved File Browser Button Tooltips
In the File Browser the "Cancel" button always has a tooltip hint of
"Cancel loading of selected file", even when you are saving. This PR
changes that to a more generic "Cancel file operation". Similarly the
save/load button always shows "Execute selected file". This PR instead
uses the description of the operator that is called when that button is
pressed. When opening a file it shows "Open a Blender file", saving
shows "Save the current file in the desired location", etc.