Fix #130861: Crash when displaying asset shelf with 1000s of assets
Once again the issue is items from the file list cache that get freed, once the
cache exceeds its maximum size. Although we already minimized use of these
items, we'd still keep a reference to all assets to be displayed in the shelf
for the active catalog.

Avoid this entirely, and only request the items from cache for the few assets
visible on screen (excludes assets not visible because of scrolling). This way
we only use the cached item briefly and don't have to store it in any way.

Note that the items from the file list cache are only needed for loading the
previews. #122439 would change this, so it's finally independed of the cache.