Anim: make `Slot::users_remove()` gracefully deal with duplicate users
`animrig::Slot::users_remove()` now also works when there are multiple
occurrences of the same ID. Because of the above-mentioned pointer
remapping now working correctly, the pointers in the slot user list
are not 100% under control of the Slot class, and thus there could
hypothetically be duplicates there.

This doesn't fix any concrete issues, it's just a safety measure.