VFONT: Text Objects Fallback Only When Default Font
With Text Objects, if the default vfont font is selected (usually
because nothing has been selected) then allow use of the fallback font
stack when characters are not found in it. This allows a nice first
experience, for testing, and allows the translation of the initial
"Text" string.  However, if the user selects a different specific font
then do not use fallback, only allow characters that are included in
that font. This should allow finer control of what is shown when
selecting specific fonts, and a more consistent output if we change or
update the UI fonts.