Docs: Misplaced En Dash in Main README
In American English, I think this should be a long dash (called an em dash), not a short dash/hyphen (called an en dash). However, for a global audience, switching to the colon may be better.

Actual Problem: The original short dash makes the joint between thoughts smoosh together in a way that's hard to read, it really needs more white space to separate the thoughts.

The dictionary company's article, below, says, "Like commas and parentheses, em dashes set off extra information, such as examples" — I would argue that dash in the sentence I'm asking to change is a dash being used to set off extra information, such as examples (of facets of the 3D pipeline). Therefore, this falls into the category of long dash/en dash or colon, not short dash/en dash.

https://www.merriam-webster.com/grammar/em-dash-en-dash-how-to-use