Fix #131880: Multires & Front Faces option can cause artifacts
Introduced in 750e3a6fbaead58d931ef159d0edf4033c624d5d

The commit above incorrectly passes the `grids` `Span` into the
`calc_front_face` function, causing the calculation to miss vertices in
certain cases, as the function attempts to index instead of reading the
`normals` `Span` directly.

This has the potential to cause crashes too, as the indices of the grid
are not guaranteed to be bounded by the size of the `factors` array.