ImBuf API: support for loading images from memory
Previously in my addons I had to create temporary files
to load an image from a zip archive, which was a bottleneck.
Now there was a need to speed it up. Loading an image
from bytes seemed like a good solution for this.
I found a discussion where this option was proposed:
https://devtalk.blender.org/t/image-python-api-for-blender/174/9

Refers to #54272